使用 gorm 自带的数据库迁移
对于列或者索引只会增不会减，会保留所有数据

+++ src/main/go/com.github.mushanwb.gin_web/model/base_model.go
@@ -2,7 +2,7 @@ package model
 
 type BaseModel struct {
 	ID        int64 `gorm:"primaryKey;autoIncrement;not null" json:"id"`
-	CreatedAt int64 `json:"created_at"`
-	UpdatedAt int64 `json:"updated_at"`
-	DeletedAt int64 `json:"-"`
+	CreatedAt int32 `json:"created_at"`
+	UpdatedAt int32 `json:"updated_at"`
+	DeletedAt int32 `json:"-"`
 }
